requirements.txt
google-adk>=1.27.0
google-genai>=1.68.0
google-cloud-aiplatform>=1.40.0
atlassian-python-api>=3.41.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
